site.yml: Adjust paths to match specifications
According to the Antora docs the url: key should point to the git
repo, and start_path should be the path from the git repo to the
component. Although combining these into one path have worked so
far, it is likely we will run into trouble later. One example
is the edit_url, which uses start_path and assumes it is as
documented.